requirements.txt
fastapi
uvicorn[standard]
openai
psycopg2-binary
redis
celery
feedparser
discord.py
imapclient
